node cache tests full; fix returns in put func

+++ internal/packages/nodes/modules/verifier/EngineVerifier.go
@@ -12,7 +12,7 @@ type EngineVerifier struct{}
 // Engine ruleset
 const SECONDS_IN_MS = 1000
 const MAX_BLOCKS_TRAVEL_PER_SECOND = 4
-const MAX_VEL = MAX_BLOCKS_TRAVEL_PER_SECOND / SECONDS_IN_MS
+const MAX_VEL = float64(MAX_BLOCKS_TRAVEL_PER_SECOND) / float64(SECONDS_IN_MS)
 
 // Engine verify functions
 
@@ -26,10 +26,14 @@ func (ev *EngineVerifier) VerifyPlayerMovement(startPkg *ActionDataPackageTimest
 	blocksTraveledY := math.Abs(float64(endPos.Y) - float64(startPos.Y))
 	blocksTraveledTotal := blocksTraveledX + blocksTraveledY
 
-	velocity := int(blocksTraveledTotal) / elapsedTime
+	velocity := blocksTraveledTotal / float64(elapsedTime)
+
+	// DEBUG
+	// fmt.Println("VELOCITY: ", velocity)
+	// fmt.Println("MAX VELOCITY: ", MAX_VEL)
 
 	if velocity > MAX_VEL {
-		fmt.Printf("[ERR] Engine player movement verification (velocity too high)\nVelocity: %d, max acceptable velocity: %d", velocity, MAX_VEL)
+		fmt.Printf("[ERR] Engine player movement verification (velocity too high)\nVelocity: %f, max acceptable velocity: %f", velocity, MAX_VEL)
 		return false
 	}
